+++ doc/source/serve/doc_code/gradio-integration-parallel.py
@@ -52,3 +52,17 @@ class MyGradioServer(GradioIngress):
 # __doc_app_begin__
 app = MyGradioServer.bind(app1, app2)
 # __doc_app_end__
+
+# Test example code
+import requests
+
+serve.run(app)
+response = requests.post(
+    "http://127.0.0.1:8000/api/predict/", json={"data": ["My name is Lewis"]}
+)
+assert response.status_code == 200
+print(
+    "gradio-integration-parallel.py: Response from example code is", 
+    response.json()["data"]
+)
+serve.shutdown()
