+++ doc/source/serve/doc_code/gradio-original.py
@@ -21,5 +21,14 @@ def model2(text):
 demo = gr.Interface(
     lambda text: f"{model1(text)}\n------------\n{model2(text)}", "textbox", "textbox"
 )
-demo.launch()
 # __doc_code_end__
+
+# Test example code
+import requests
+
+demo.launch(prevent_thread_lock=True)
+response = requests.post(
+    "http://127.0.0.1:7860/api/predict/", json={"data": ["My name is Lewis"]}
+)
+assert response.status_code == 200
+print("gradio-original.py: Response from example code is", response.json()["data"])
